Move helper functions to a separate file
Refactor the code to have the helper functions in a separate file for
easier readability.
